requirements.txt
fastapi
pydantic
typing
langchain_community
langgraph
langchain_groq
python-dotenv
uvicorn
streamlit
langchain_community 
langsmith 
langgraph
langchain-openai